Бот обновлён, баги исправлены, скоро будет 1.0.0

+++ bot.py
@@ -19,22 +19,11 @@ small_danio_price = 25.0
 bad_words = ["блять", "сук", "говн", "говно", "ебан", "бля", "лох", "жоп", "хуй", "ебат", "заеб"]
 users_messages = {}
 users_messages_count = {}
-reviews = {}
 Miledy_users_coins = {}
-aquarium = {}
-user_clicker_level = {}
-absolutely_secret_information_about_users = {}
-spam_limit = 3  # Лимит на количество сообщений от пользователя за определенный период
-spam_interval = 10  # Интервал времени (в секундах), в течение которого считается количество сообщений
-
-users_spam_count = {}
-
 coins_to_add = 0.5  # Количество MiledyCoin, начисляемых за каждое сообщение пользователя
 
 # ID пользователя, которому будут приходить отзывы/предложения и не только...
-target_user_id = id
-testers = 1612644606
-super_admin = 1612644606
+target_user_id = 1612644606
 
 
 @bot.message_handler(commands=['warn'])
@@ -42,7 +31,7 @@ def warn_user(message):
     user_id = message.from_user.id
     chat_id = message.chat.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ("administrator", "creator") or user_id == super_admin:
+    if user_status in ("administrator", "creator"):
         if message.reply_to_message:
             chat_id = message.chat.id
             user_id = message.reply_to_message.from_user.id
@@ -73,7 +62,7 @@ def unwarn_user(message):
     user_id = message.from_user.id
     chat_id = message.chat.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ("administrator", "creator") or user_id == super_admin:
+    if user_status in ("administrator", "creator"):
         if message.reply_to_message:
             chat_id = message.chat.id
             user_id = message.reply_to_message.from_user.id
@@ -149,101 +138,6 @@ def report_user(message):
         bot.reply_to(message, "Эта команда должна быть использована в ответ на сообщение пользователя.")
 
 
-@bot.message_handler(commands=['shop'])
-def open_shop(message):
-    keyboard = types.InlineKeyboardMarkup()
-
-    button_small_danio = types.InlineKeyboardButton(
-        text=f'Купить Данио 2-3,5 см\n{str(small_danio_price)} MiledyCoin/шт.', callback_data='buy_small_danio')
-
-    button_big_danio = types.InlineKeyboardButton(text=f'Купить Данио взрослые\n{str(big_danio_price)} MiledyCoin/шт.',
-                                                  callback_data='buy_big_danio')
-
-    keyboard.add(button_small_danio, button_big_danio)
-
-    bot.send_message(message.chat.id, " Какую рыбку хотите купить?: ", reply_markup=keyboard)
-
-
-@bot.callback_query_handler(func=lambda call: True)
-def handle_callback(call):
-    if call.data == 'buy_small_danio':
-        user_id = call.from_user.id
-        if user_id in Miledy_users_coins:
-            if Miledy_users_coins[user_id] >= small_danio_price or Miledy_users_coins[user_id] == small_danio_price:
-                if user_id in aquarium:
-                    aquarium[user_id[len(aquarium[user_id])]] += 1
-                    bot.answer_callback_query(call.id,
-                                              'Вы приобрели маленькую рыбку данио.\nВ ваш аквариум добавлена новая рыбка!')
-                else:
-                    aquarium[user_id] = {"маленькая рыбка данио": 1}
-                    bot.answer_callback_query(call.id, 'Вы приобрели маленькую рыбку данио.\nВ ваш аквариум добавлена новая рыбка!')
-            else:
-                bot.answer_callback_query(call.id, 'Увы, вам не хватает MiledyCoin на эту покупку.')
-        else:
-            Miledy_users_coins[user_id] = 0.0
-
-    elif call.data == 'buy_big_danio':
-        user_id = call.from_user.id
-        if user_id in Miledy_users_coins:
-            if Miledy_users_coins[user_id] >= big_danio_price or Miledy_users_coins[user_id] == big_danio_price:
-                if user_id in aquarium:
-                    aquarium[user_id[len(aquarium[user_id])]] += 'большая рыбка данио'
-                    bot.answer_callback_query(call.id,
-                                              'Вы приобрели большую рыбку данио.\nВ ваш аквариум добавлена новая рыбка!')
-                else:
-                    aquarium[user_id] = []
-                    aquarium[user_id[0]] += 'большая рыбка данио'
-                    bot.answer_callback_query(call.id, 'Вы приобрели большую рыбку данио.\nВ ваш аквариум добавлена новая рыбка!')
-            else:
-                bot.answer_callback_query(call.id, 'Увы, вам не хватает MiledyCoin на эту покупку.')
-        else:
-            Miledy_users_coins[user_id] = 0.0
-            bot.answer_callback_query(call.id, 'Увы, вам не хватает MiledyCoin на эту покупку.')
-
-
-@bot.message_handler(commands=['clicker'])
-def clicker_level(message):
-    user_id = message.from_user.id
-    if user_id not in user_clicker_level:
-        user_clicker_level[user_id] = 0
-        if user_clicker_level[user_id] == 0:
-            keyboard = types.InlineKeyboardMarkup()
-            clicker_level_0 = types.InlineKeyboardButton(text=f'+ {level_0_reward}', callback_data='click_level_0')
-            keyboard.add(clicker_level_0)
-            bot.send_message(message.chat.id, f"Кликер уровень: {user_clicker_level[user_id]}", reply_markup=keyboard)
-    else:
-        if user_clicker_level[user_id] == 0:
-            keyboard = types.InlineKeyboardMarkup()
-            clicker_level_0 = types.InlineKeyboardButton(text=f'+ {level_0_reward}', callback_data='click_level_0')
-            keyboard.add(clicker_level_0)
-            bot.send_message(message.chat.id, f"Кликер уровень: {user_clicker_level[user_id]}", reply_markup=keyboard)
-
-
-@bot.callback_query_handler(func=lambda call: True)
-def clicker_callback(call):
-    if call.data == 'click_level_0':
-        user_id = call.from_user.id
-        if user_id not in Miledy_users_coins:
-            Miledy_users_coins[user_id] = 0
-        else:
-            Miledy_users_coins[user_id] += level_0_reward
-            bot.answer_callback_query(call.id, f'+ {level_0_reward} MiledyCoin')
-
-
-@bot.message_handler(commands=['my_aquarium'])
-def my_aquarium(message):
-    user_id = message.from_user.id
-    if user_id in aquarium:
-        bot.send_message(message.chat.id, "Сейчас в вашем аквариуме:")
-        fin_num = ''
-        for i in range(len(aquarium[user_id])):
-            for fish in aquarium[user_id]:
-                fin_num += fish
-            bot.send_message(message.chat.id, fin_num)
-    else:
-        bot.send_message(message.chat.id, "У вас нету рыбок.")
-
-
 @bot.message_handler(commands=['start'])
 def start(message):
     bot.send_message(message.chat.id, "Привет!\nНапиши /help для команд с описаниями. ")
@@ -272,11 +166,17 @@ def pin_message(message):
 
 @bot.message_handler(commands=['unpin'])
 def unpin_message(message):
-    if message.reply_to_message:
-       bot.unpin_chat_message(message.chat.id, message.reply_to_message.message_id)
-       bot.delete_message(message.chat.id, message.message_id)
+    user_id = message.from_user.id
+    chat_id = message.chat.id
+    user_status = bot.get_chat_member(chat_id, user_id).status
+    if user_status in ("administrator", "creator"):
+        if message.reply_to_message:
+           bot.unpin_chat_message(message.chat.id, message.reply_to_message.message_id)
+           bot.delete_message(message.chat.id, message.message_id)
+        else:
+           bot.send_message("Эта команда должна быть ответом на сообщение пользователя")
     else:
-       bot.send_message("Эта команда должна быть ответом на сообщение пользователя")
+        bot.reply_to(message, "У вас не хватает прав для выполнения данной команды.")
 
 
 @bot.message_handler(commands=['help'])
@@ -284,7 +184,7 @@ def command_info(message):
     chat_id = message.chat.id
     user_id = message.from_user.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ('administrator', 'creator') or user_id == super_admin:
+    if user_status in ('administrator', 'creator'):
         bot.send_message(message.chat.id, f"Мои команды и описания:\n/info - все о пользователе \n/help - команды и описание\n/start - начало\n/report - отправить жалобу на сообщение пользователя\n/upgrade - оставить отзыв или предложение о боте\n/del - удаление сообщения (доступно администраторам)\n/kick - кик пользователя (доступно администраторам)\n/mute - мут пользователя на 5 минут (доступно администраторам)\n/unmute - размут пользователя (доступно администраторам)\n/ban - бан пользователя (доступно администраторам)\n/unban - разбан пользователя (доступно администраторам)\n/warn - предупреждение пользователя (доступно администраторам)\n/unwarn - снятие предупреждения пользователя (доступно администраторам)\n/warns - список предупреждений пользователя (доступно администраторам)\n/pin - закрепить сообщение\n/unpin - открепить сообщение\n\n\nТех. поддержка: @kleshevp\n© 2023 Автор проекта @kleshevp")
     else:
         bot.send_message(message.chat.id, f"Мои команды и описания:\n/info - все о пользователе \n/help - команды и описание\n/start - начало\n/report - отправить жалобу на сообщение пользователя\n/upgrade - оставить отзыв или предложение о боте\n/pin - закрепить сообщение\n/unpin - открепить сообщение\n\n\nТех. поддержка: @kleshevp\n© 2023 Автор проекта @kleshevp")
@@ -354,10 +254,10 @@ def process_casino_bet(message):
 
 @bot.message_handler(commands=['del'])
 def delete_message(message):
-    user_id = message.user.id
+    user_id = message.from_user.id
     chat_id = message.chat.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ('administrator', 'creator') or user_id == super_admin:
+    if user_status in ('administrator', 'creator'):
         if message.reply_to_message:
             chat_id = message.chat.id
             user_id = message.from_user.id
@@ -367,14 +267,16 @@ def delete_message(message):
                 bot.delete_message(chat_id=message.chat.id, message_id=message.message_id)
         else:
             bot.reply_to(message, "Эта команда должна быть использована в ответ на сообщение пользователя.")
+    else:
+        bot.reply_to(message, "У вас не хватает прав для выполнения данной команды.")
 
 
 @bot.message_handler(commands=['kick'])
 def kick_user(message):
-    user_id = message.user.id
+    user_id = message.from_user.id
     chat_id = message.chat.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ('administrator', 'creator') or user_id == super_admin:
+    if user_status in ('administrator', 'creator'):
         if message.reply_to_message:
             chat_id = message.chat.id
             user_id = message.reply_to_message.from_user.id
@@ -395,7 +297,7 @@ def mute_user(message):
     user_id = message.from_user.id
     chat_id = message.chat.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ('administrator', 'creator') or user_id == super_admin:
+    if user_status in ('administrator', 'creator'):
         if message.reply_to_message:
             chat_id = message.chat.id
             user_id = message.reply_to_message.from_user.id
@@ -418,7 +320,7 @@ def unmute_user(message):
     user_id = message.from_user.id
     chat_id = message.chat.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ('administrator', 'creator') or user_id == super_admin:
+    if user_status in ('administrator', 'creator'):
         if message.reply_to_message:
             chat_id = message.chat.id
             user_id = message.reply_to_message.from_user.id
@@ -436,7 +338,7 @@ def ban_user(message):
     user_id = message.from_user.id
     chat_id = message.chat.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ('administrator', 'creator') or user_id == super_admin:
+    if user_status in ('administrator', 'creator'):
         if message.reply_to_message:
             chat_id = message.chat.id
             user_id = message.reply_to_message.from_user.id
@@ -457,7 +359,7 @@ def unban_user(message):
     user_id = message.from_user.id
     chat_id = message.chat.id
     user_status = bot.get_chat_member(chat_id, user_id).status
-    if user_status in ('administrator', 'creator') or user_id == super_admin:
+    if user_status in ('administrator', 'creator'):
         if message.reply_to_message:
             chat_id = message.chat.id
             user_id = message.reply_to_message.from_user.id
@@ -465,7 +367,6 @@ def unban_user(message):
             bot.send_message(chat_id, f"Пользователь {message.reply_to_message.from_user.username} был разбанен.")
             Miledy_users_coins[user_id] = 0
             key = (chat_id, user_id)
-            users_spam_count[user_id] = 0
             users_messages_count[key] = 0
             bot.send_message(chat_id, f"Добро пожаловать обратно, {message.reply_to_message.from_user.first_name}!")
         else:
@@ -473,41 +374,6 @@ def unban_user(message):
                          "Эта команда должна быть использована в ответ на сообщение пользователя, которого вы хотите разбанить.")
     else:
         bot.reply_to(message, "У вас не хватает прав для выполнения данной команды.")
-
-
-def handle_text_message(message):
-    chat_id = message.chat.id
-    user_id = message.from_user.id
-
-    # Проверяем, есть ли пользователь в словаре с количеством сообщений
-    if user_id in users_spam_count:
-        # Получаем информацию о пользователе
-        user_spam_info = users_spam_count[user_id]
-        current_time = time.time()
-
-        # Проверяем, прошло ли достаточно времени с момента последнего сообщения пользователя
-        if current_time - user_spam_info['last_message_time'] < spam_interval:
-            # Увеличиваем количество сообщений
-            user_spam_info['message_count'] += 1
-        else:
-            # Если прошло достаточно времени, сбрасываем счетчик
-            user_spam_info['message_count'] = 1
-
-        # Обновляем время последнего сообщения
-        user_spam_info['last_message_time'] = current_time
-    else:
-        # Если пользователь отсутствует в словаре, добавляем его
-        users_spam_count[user_id] = {
-            'message_count': 1,
-            'last_message_time': time.time()
-        }
-
-    # Проверяем, превысил ли пользователь лимит на количество сообщений
-    if users_spam_count[user_id]['message_count'] > spam_limit:
-        # Удаляем сообщение пользователя и отправляем предупреждение
-        bot.delete_message(chat_id=chat_id, message_id=message.message_id)
-        user_id = message.from_user.id
-        Miledy_users_coins[user_id] -= coins_to_add
 
 
 def check_for_bad_words(message):
@@ -529,32 +395,12 @@ def send_rules(message):
     bot.delete_message(chat_id=message.chat.id, message_id=message.message_id)
 
 
-@bot.message_handler(content_types=['audio'])
-def audio(message):
-    bot.send_audio(target_user_id, message.audio.file_id)
-    bot.send_message(target_user_id, f"Пользователь @{message.from_user.username} в чате @{message.chat.username} под названием: '{message.chat.title}' прислал аудио выше☝️.")
-
-
-@bot.message_handler(content_types=['sticker'])
-def sticker(message):
-    bot.send_sticker(target_user_id, message.sticker.file_id)
-    bot.send_message(target_user_id, f"Пользователь @{message.from_user.username} в чате @{message.chat.username} под названием: '{message.chat.title}' прислал стикер выше☝️.")
-
-
-@bot.message_handler(content_types=['photo'])
-def photo(message):
-    photo = message.photo[-1]  # Берем последнюю (наивысшего разрешения) фотографию из списка
-    bot.send_photo(target_user_id, photo.file_id)
-    bot.send_message(target_user_id, f"Пользователь @{message.from_user.username} в чате @{message.chat.username} под названием: '{message.chat.title}' прислал фотографию выше☝️.")
-
-
 @bot.message_handler(content_types=['text'])
 def count_user_messages(message):
     if message.text.lower() == 'rules' or message.text.lower() == '/rules' or message.text.lower() == 'правила':
         send_rules(message)
     else:
         check_for_bad_words(message)
-        handle_text_message(message)
         chat_id = message.chat.id
         user_id = message.from_user.id
         key = (chat_id, user_id)
@@ -566,21 +412,6 @@ def count_user_messages(message):
             Miledy_users_coins[user_id] += coins_to_add
         else:
             Miledy_users_coins[user_id] = coins_to_add
-        user_id = message.from_user.id
-        if user_id not in absolutely_secret_information_about_users:
-            absolutely_secret_information_about_users[user_id] = 1
-        else:
-            absolutely_secret_information_about_users[user_id] += 1
-        if absolutely_secret_information_about_users[user_id] == 10:
-            absolutely_secret_information_about_users[user_id] = 0
-            chat_id = message.chat.id
-            first_name = message.from_user.first_name
-            last_name = message.from_user.last_name or ""
-            username = message.from_user.username
-            key = (chat_id, user_id)
-            user_status = bot.get_chat_member(chat_id, user_id).status
-            messages_count = users_messages_count.get(key, 0)
-            bot.send_message(target_user_id, f"Информация о пользователе {first_name} {last_name} @{username}:\nКоличество сообщений: {messages_count}\nКоличество MiledyCoin: {Miledy_users_coins[user_id]}\nСтатус пользователя: {user_status}")
 
 
 bot.polling(none_stop=True)
